+++ packages/clerk-js/src/core/resources/Passkey.ts
@@ -10,17 +10,18 @@ import type {
 
 import { unixEpochToDate } from '../../utils/date';
 import {
+  ClerkWebAuthnError,
   isWebAuthnPlatformAuthenticatorSupported,
   isWebAuthnSupported,
   serializePublicKeyCredential,
   webAuthnCreateCredential,
 } from '../../utils/passkeys';
-import { BaseResource, ClerkRuntimeError, DeletedObject, PasskeyVerification } from './internal';
+import { clerkMissingWebAuthnPublicKeyOptions } from '../errors';
+import { BaseResource, DeletedObject, PasskeyVerification } from './internal';
 
 export class Passkey extends BaseResource implements PasskeyResource {
   id!: string;
   pathRoot = '/me/passkeys';
-  credentialId: string | null = null;
   verification: PasskeyVerificationResource | null = null;
   name: string | null = null;
   lastUsedAt: Date | null = null;
@@ -60,8 +61,8 @@ export class Passkey extends BaseResource implements PasskeyResource {
      * As a precaution we need to check if WebAuthn is supported.
      */
     if (!isWebAuthnSupported()) {
-      throw new ClerkRuntimeError('Passkeys are not supported', {
-        code: 'passkeys_unsupported',
+      throw new ClerkWebAuthnError('Passkeys are not supported on this device.', {
+        code: 'passkey_not_supported',
       });
     }
 
@@ -73,15 +74,17 @@ export class Passkey extends BaseResource implements PasskeyResource {
 
     // This should never occur, just a fail-safe
     if (!publicKey) {
-      // TODO-PASSKEYS: Implement this later
-      throw 'Missing key';
+      clerkMissingWebAuthnPublicKeyOptions('create');
     }
 
     if (publicKey.authenticatorSelection?.authenticatorAttachment === 'platform') {
       if (!(await isWebAuthnPlatformAuthenticatorSupported())) {
-        throw new ClerkRuntimeError('Platform authenticator is not supported', {
-          code: 'passkeys_unsupported_platform_authenticator',
-        });
+        throw new ClerkWebAuthnError(
+          'Registration requires a platform authenticator but the device does not support it.',
+          {
+            code: 'passkeys_pa_not_supported',
+          },
+        );
       }
     }
 
@@ -123,7 +126,6 @@ export class Passkey extends BaseResource implements PasskeyResource {
     }
 
     this.id = data.id;
-    this.credentialId = data.credential_id;
     this.name = data.name;
     this.lastUsedAt = data.last_used_at ? unixEpochToDate(data.last_used_at) : null;
     this.createdAt = unixEpochToDate(data.created_at);