+++ packages/clerk-js/src/ui/components/SignIn/__tests__/SignInFactorOne.test.tsx
@@ -4,7 +4,7 @@ import { describe, it, jest } from '@jest/globals';
 import { waitFor } from '@testing-library/dom';
 
 import { ClerkAPIResponseError } from '../../../../core/resources';
-import { act, render, screen } from '../../../../testUtils';
+import { act, mockWebAuthn, render, screen } from '../../../../testUtils';
 import { bindCreateFixtures } from '../../../utils/test/createFixtures';
 import { runFakeTimers } from '../../../utils/test/runFakeTimers';
 import { SignInFactorOne } from '../SignInFactorOne';
@@ -228,26 +228,150 @@ describe('SignInFactorOne', () => {
         });
       });
 
-      it('redirects to `reset-password` if signIn requires a new password', async () => {
+      it('Prompts the user to reset their password via email if it has been pwned', async () => {
         const { wrapper, fixtures } = await createFixtures(f => {
           f.withEmailAddress();
           f.withPassword();
           f.withPreferredSignInStrategy({ strategy: 'password' });
-          f.startSignInWithPhoneNumber({ supportPassword: true });
+          f.startSignInWithEmailAddress({
+            supportPassword: true,
+            supportEmailCode: true,
+            supportResetPassword: true,
+          });
         });
         fixtures.signIn.prepareFirstFactor.mockReturnValueOnce(Promise.resolve({} as SignInResource));
 
-        fixtures.signIn.attemptFirstFactor.mockReturnValueOnce(
-          Promise.resolve({ status: 'needs_new_password' } as SignInResource),
+        const errJSON = {
+          code: 'form_password_pwned',
+          long_message:
+            'Password has been found in an online data breach. For account safety, please reset your password.',
+          message: 'Password has been found in an online data breach. For account safety, please reset your password.',
+          meta: { param_name: 'password' },
+        };
+
+        fixtures.signIn.attemptFirstFactor.mockRejectedValueOnce(
+          new ClerkAPIResponseError('Error', {
+            data: [errJSON],
+            status: 422,
+          }),
         );
 
         await runFakeTimers(async () => {
           const { userEvent } = render(<SignInFactorOne />, { wrapper });
           await userEvent.type(screen.getByLabelText('Password'), '123456');
           await userEvent.click(screen.getByText('Continue'));
+
           await waitFor(() => {
-            expect(fixtures.router.navigate).toHaveBeenCalledWith('../reset-password');
+            screen.getByText('Password compromised');
+            screen.getByText(
+              'This password has been found as part of a breach and can not be used, please reset your password.',
+            );
+            screen.getByText('Or, sign in with another method');
           });
+
+          await userEvent.click(screen.getByText('Reset your password'));
+          screen.getByText('First, enter the code sent to your email ID');
+        });
+      });
+
+      it('Prompts the user to reset their password via phone if it has been pwned', async () => {
+        const { wrapper, fixtures } = await createFixtures(f => {
+          f.withEmailAddress();
+          f.withPassword();
+          f.withPreferredSignInStrategy({ strategy: 'password' });
+          f.startSignInWithPhoneNumber({
+            supportPassword: true,
+            supportPhoneCode: true,
+            supportResetPassword: true,
+          });
+        });
+        fixtures.signIn.prepareFirstFactor.mockReturnValueOnce(Promise.resolve({} as SignInResource));
+
+        const errJSON = {
+          code: 'form_password_pwned',
+          long_message:
+            'Password has been found in an online data breach. For account safety, please reset your password.',
+          message: 'Password has been found in an online data breach. For account safety, please reset your password.',
+          meta: { param_name: 'password' },
+        };
+
+        fixtures.signIn.attemptFirstFactor.mockRejectedValueOnce(
+          new ClerkAPIResponseError('Error', {
+            data: [errJSON],
+            status: 422,
+          }),
+        );
+
+        await runFakeTimers(async () => {
+          const { userEvent } = render(<SignInFactorOne />, { wrapper });
+          await userEvent.type(screen.getByLabelText('Password'), '123456');
+          await userEvent.click(screen.getByText('Continue'));
+
+          await waitFor(() => {
+            screen.getByText('Password compromised');
+            screen.getByText(
+              'This password has been found as part of a breach and can not be used, please reset your password.',
+            );
+            screen.getByText('Or, sign in with another method');
+          });
+
+          await userEvent.click(screen.getByText('Reset your password'));
+          screen.getByText('First, enter the code sent to your phone');
+        });
+      });
+
+      it('entering a pwned password, then going back and clicking forgot password should result in the correct title', async () => {
+        const { wrapper, fixtures } = await createFixtures(f => {
+          f.withEmailAddress();
+          f.withPassword();
+          f.withPreferredSignInStrategy({ strategy: 'password' });
+          f.startSignInWithEmailAddress({
+            supportPassword: true,
+            supportEmailCode: true,
+            supportResetPassword: true,
+          });
+        });
+        fixtures.signIn.prepareFirstFactor.mockReturnValueOnce(Promise.resolve({} as SignInResource));
+
+        const errJSON = {
+          code: 'form_password_pwned',
+          long_message:
+            'Password has been found in an online data breach. For account safety, please reset your password.',
+          message: 'Password has been found in an online data breach. For account safety, please reset your password.',
+          meta: { param_name: 'password' },
+        };
+
+        fixtures.signIn.attemptFirstFactor.mockRejectedValueOnce(
+          new ClerkAPIResponseError('Error', {
+            data: [errJSON],
+            status: 422,
+          }),
+        );
+
+        await runFakeTimers(async () => {
+          const { userEvent } = render(<SignInFactorOne />, { wrapper });
+          await userEvent.type(screen.getByLabelText('Password'), '123456');
+          await userEvent.click(screen.getByText('Continue'));
+
+          await waitFor(() => {
+            screen.getByText('Password compromised');
+            screen.getByText(
+              'This password has been found as part of a breach and can not be used, please reset your password.',
+            );
+            screen.getByText('Or, sign in with another method');
+          });
+
+          // Go back
+          await userEvent.click(screen.getByText('Back'));
+
+          // Choose to reset password via "Forgot password" instead
+          await userEvent.click(screen.getByText(/Forgot password/i));
+          screen.getByText('Forgot Password?');
+          expect(
+            screen.queryByText(
+              'This password has been found as part of a breach and can not be used, please reset your password.',
+            ),
+          ).not.toBeInTheDocument();
         });
       });
     });
@@ -601,6 +725,59 @@ describe('SignInFactorOne', () => {
         });
       });
     });
+
+    describe('Passkey', () => {
+      it('shows the next available factor because webauthn is not supported', async () => {
+        const { wrapper, fixtures } = await createFixtures(f => {
+          f.withEmailAddress();
+          f.withPassword();
+          f.withPasskey();
+          f.withPreferredSignInStrategy({ strategy: 'otp' });
+          f.startSignInWithEmailAddress({ supportPasskey: true, supportEmailCode: true });
+        });
+        fixtures.signIn.prepareFirstFactor.mockReturnValueOnce(Promise.resolve({} as SignInResource));
+        render(<SignInFactorOne />, { wrapper });
+        screen.getByText('Check your email');
+      });
+
+      mockWebAuthn(() => {
+        it('shows a passkey factor one screen', async () => {
+          const { wrapper } = await createFixtures(f => {
+            f.withEmailAddress();
+            f.withPassword();
+            f.withPasskey();
+            f.withPreferredSignInStrategy({ strategy: 'otp' });
+            f.startSignInWithEmailAddress({ supportPasskey: true, supportEmailCode: true });
+          });
+          render(<SignInFactorOne />, { wrapper });
+          screen.getByText('Use your passkey');
+          screen.getByText(
+            "Using your passkey confirms it's you. Your device may ask for your fingerprint, face or screen lock.",
+          );
+          screen.getByText('hello@clerk.com');
+        });
+
+        it('call appropriate method from passkey factor one screen', async () => {
+          const { wrapper, fixtures } = await createFixtures(f => {
+            f.withEmailAddress();
+            f.withPassword();
+            f.withPasskey();
+            f.withPreferredSignInStrategy({ strategy: 'otp' });
+            f.startSignInWithEmailAddress({ supportPasskey: true, supportEmailCode: true });
+          });
+          fixtures.signIn.__experimental_authenticateWithPasskey.mockResolvedValue({
+            status: 'complete',
+          } as SignInResource);
+          const { userEvent } = render(<SignInFactorOne />, { wrapper });
+
+          await userEvent.click(screen.getByText('Continue'));
+
+          await waitFor(() => {
+            expect(fixtures.signIn.__experimental_authenticateWithPasskey).toHaveBeenCalled();
+          });
+        });
+      });
+    });
   });
 
   describe('Use another method', () => {
@@ -665,6 +842,43 @@ describe('SignInFactorOne', () => {
       screen.getByText(`Sign in with your password`);
       const deactivatedMethod = screen.queryByText(`Send link to ${email}`);
       expect(deactivatedMethod).not.toBeInTheDocument();
+    });
+
+    it('should skip passkey alternative method when webauthn is not supported', async () => {
+      const email = 'test@clerk.com';
+      const { wrapper, fixtures } = await createFixtures(f => {
+        f.withEmailAddress();
+        f.withPassword();
+        f.withPasskey();
+        f.withPreferredSignInStrategy({ strategy: 'otp' });
+        f.startSignInWithEmailAddress({ supportPasskey: true, supportEmailCode: true, identifier: email });
+      });
+      fixtures.signIn.prepareFirstFactor.mockReturnValueOnce(Promise.resolve({} as SignInResource));
+      const { userEvent } = render(<SignInFactorOne />, { wrapper });
+      await userEvent.click(screen.getByText('Use another method'));
+      await userEvent.click(screen.getByText('Sign in with your password'));
+      await userEvent.click(screen.getByText('Use another method'));
+      const deactivatedMethod = screen.queryByText(`Sign in with your passkey`);
+      expect(deactivatedMethod).not.toBeInTheDocument();
+    });
+
+    mockWebAuthn(() => {
+      it('should not skip passkey alternative method when webauthn is supported', async () => {
+        const email = 'test@clerk.com';
+        const { wrapper, fixtures } = await createFixtures(f => {
+          f.withEmailAddress();
+          f.withPassword();
+          f.withPasskey();
+          f.withPreferredSignInStrategy({ strategy: 'otp' });
+          f.startSignInWithEmailAddress({ supportPasskey: true, supportEmailCode: true, identifier: email });
+        });
+        fixtures.signIn.prepareFirstFactor.mockReturnValueOnce(Promise.resolve({} as SignInResource));
+        const { userEvent } = render(<SignInFactorOne />, { wrapper });
+        await userEvent.click(screen.getByText('Use another method'));
+        await userEvent.click(screen.getByText('Sign in with your password'));
+        await userEvent.click(screen.getByText('Use another method'));
+        screen.getByText(`Sign in with your passkey`);
+      });
     });
 
     it('should list enabled first factor methods without the current one', async () => {