+++ packages/clerk-js/src/ui/utils/test/fixtureHelpers.ts
@@ -99,6 +99,7 @@ const createSignInFixtureHelpers = (baseClient: ClientJSON) => {
     supportEmailCode?: boolean;
     supportEmailLink?: boolean;
     supportResetPassword?: boolean;
+    supportPasskey?: boolean;
   };
 
   type SignInWithPhoneNumberParams = {
@@ -124,12 +125,14 @@ const createSignInFixtureHelpers = (baseClient: ClientJSON) => {
       supportEmailCode,
       supportEmailLink,
       supportResetPassword,
+      supportPasskey,
     } = params || {};
     baseClient.sign_in = {
       status: 'needs_first_factor',
       identifier,
       supported_identifiers: ['email_address'],
       supported_first_factors: [
+        ...(supportPasskey ? [{ strategy: 'passkey' }] : []),
         ...(supportPassword ? [{ strategy: 'password' }] : []),
         ...(supportEmailCode ? [{ strategy: 'email_code', safe_identifier: identifier || 'n*****@clerk.com' }] : []),
         ...(supportEmailLink ? [{ strategy: 'email_link', safe_identifier: identifier || 'n*****@clerk.com' }] : []),
@@ -362,11 +365,18 @@ const createUserSettingsFixtureHelpers = (environment: EnvironmentJSON) => {
       enabled: true,
       required: false,
       used_for_first_factor: true,
-      first_factors: [],
+      first_factors: ['passkey'],
       used_for_second_factor: false,
       second_factors: [],
-      verifications: [],
+      verifications: ['passkey'],
       verify_at_sign_up: false,
+      ...opts,
+    };
+  };
+
+  const withPasskeySettings = (opts?: Partial<UserSettingsJSON['passkey_settings']>) => {
+    us.passkey_settings = {
+      ...us.passkey_settings,
       ...opts,
     };
   };
@@ -474,5 +484,6 @@ const createUserSettingsFixtureHelpers = (environment: EnvironmentJSON) => {
     withBackupCode,
     withAuthenticatorApp,
     withPasskey,
+    withPasskeySettings,
   };
 };