+++ packages/types/src/token.ts
@@ -2,6 +2,6 @@ import type { JWT } from './jwt';
 import type { ClerkResource } from './resource';
 
 export interface TokenResource extends ClerkResource {
-  jwt: JWT;
+  jwt?: JWT;
   getRawString: () => string;
 }