+++ packages/gatsby-plugin-clerk/CHANGELOG.md
@@ -1,5 +1,29 @@
 # Change Log
 
+## 5.0.0-beta.38
+
+### Patch Changes
+
+- Updated dependencies [[`988a299c0`](https://github.com/clerk/javascript/commit/988a299c0abf708e905592c29e394f8e4d79968e)]:
+  - @clerk/backend@1.0.0-beta.31
+  - @clerk/clerk-sdk-node@5.0.0-beta.32
+
+## 5.0.0-beta.37
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @clerk/backend@1.0.0-beta.30
+  - @clerk/clerk-react@5.0.0-beta.33
+  - @clerk/clerk-sdk-node@5.0.0-beta.31
+
+## 5.0.0-beta.36
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @clerk/clerk-react@5.0.0-beta.32
+
 ## 5.0.0-beta.35
 
 ### Patch Changes